examples/requirements.txt
paho_mqtt==1.6.1
PyYAML==6.0
